Added environmental variable PYFLUENT_ADDITIONAL_ARGUMENTS
Added an environmental variable to allow a user to pass additional
arguments to the launcher without having to change their python file.

Launcher checks if the environment variable is set and that the
user did not launch with the additional_arguments option.
If this is the case, the user additional_arguments are used instead.