refactor: field data transaction (#3819)
New design supports the following:

1. Creating a request object which can be passed for adding to a
transaction or getting the field data.
For example, for adding and getting a surface data:
```Python
>>> surface_request = SurfaceFieldDataRequest(
>>>         surfaces=[1, "hot-inlet"],
>>>         data_types=[SurfaceDataType.Vertices, SurfaceDataType.FacesCentroid],
>>>     )

>>> transaction = transaction.add_requests(surface_request)

>>> data = transaction.get_response()

>>> surface_data = data.get_field_data(surface_request)
```
Or, directly get the field data:
```Python
>>> surface_data = field_data.get_field_data(surface_request)
```
2. The request objects are NamedTuples. So, for creating a new object by
just updating a few parameters one can just use the "_replace"
```Python
>>> surface_request = surface_request._replace(surfaces=[3])
```
3. For transaction one can add requests by chaining or by adding
multiple requests together:
```Python
>>> data = transaction.add_requests(req1).add_requests(req2).add_requests(req3).add_requests(req4, req5, req6).get_response()
```
4. In summary you can just form the request objects and use the
"get_field_data" for fetching the field data from both server as well as
from a transaction and "add_requests" to add data to transactions.

**NOTE:**
The earlier methods are still usable but with a deprecation warning.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>