fix: Fix performance issues while settings boundary condition (#4166)
There were 111 gRPC calls for each `set_state` at a boundary condition
object. Those huge number of calls are due to the following 2 issues:
1. `_check_stable()` makes a number of `active?` and
`webui-release-active?` attribute queries.
2. There is an unintentional call to `boundary_conditions.__len__`
(->`_ChildNamedObjectAccessorMixin.__len__`) which further calls
`active?` and `get_object_names()` for each boundary types.

I have cleaned up the code related to
`_check_stable()`/`webui-release-active?` as we are not planning to
follow that approach for PyFluent (current plan -
https://github.com/ansys/pyfluent/issues/4128). Issue 2 is also fixed.

With this PR, there are now 4 gRPC calls for each `set_state` which is
the expected number. The runtime for the defect case has been reduced
from 12 sec. to 0.7 sec.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>