fix: Fix an issue in returning parameter units (#4177)
We retrieve the Fluent unit strings via scheme calls from PyFluent at
the end of the execution of `parameters.output_parameters.list()`
command in PyFluent so we can populate Fluent unit strings in that
command's return value. This PR prevents a Fluent-side error which was
coming from those scheme calls.

`parameters.output_parameters.list()` will be fixed in 26.1 to contain
the Fluent unit strings in the return value. The above PyFluent
workaround will still be needed in Fluent versions <= 25.2.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>