fix: nightly optislang integration tests (#4189)
Closes #4186

Tests were assuming a standalone Fluent session, or relying on the
kludge that paths could be the same inside and outside containers (only
works when both are Linux and running similar systems). With the changes
suggested here, they should now work for Fluent container as well as
standalone launches in both Windows and Linux.

The fix is using relative paths while making sure that the Fluent
standalone launch is going into the same host directory that the
container launches are going to (in this case `pyfluent.EXAMPLES_PATH`).
That is where the files are being written to in the tests.

I also noticed `test_generate_read_mesh` was using TUI commands, so I
included changes to have it use the meshing datamodel commands instead

[Test Run Dev Version
Nightly](https://github.com/ansys/pyfluent/actions/runs/15738542821) -
100% passed

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>