feat: Set string settings with allowed values via string constants (#4190)
The enumerated values are exposed as constant attributes of the
instance. Please see the unittest for usage.

There is already an attribute `allowed-values-with-display-text` in
settings API to fetch all possible allowed values which has been added
to the static-info (Fluent PR 599447).

Below is the fragment of generated class:
```
class model_2(String, AllowedValuesMixin):
    """
    Viscous Model.
    """
    _version = '261'
    fluent_name = 'model'
    _python_name = 'model'
    LAMINAR = _FlConstant('laminar')
    MIXING_LENGTH = _FlConstant('mixing-length')
    KE1E = _FlConstant('ke1e')
    INVISCID = _FlConstant('inviscid')
...
```

TODO: doc update

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>