fix: ui_mode key error in Ansys lab (#4319)
This pull request introduces a new `ui_mode` parameter in the
`PIMLauncher` class to allow specifying the user interface mode for
Fluent. The changes include updates to the class constructor, parameter
documentation, and imports.

### Addition of `ui_mode` parameter:

*
[`src/ansys/fluent/core/launcher/pim_launcher.py`](diffhunk://#diff-f3eafcfd5d76de366b9e3a94065320abf7c555ddd71c6fc2a1b6542ab8fd330bR50):
Added `UIMode` to the imports to support the new parameter.
* `class PIMLauncher` in
`src/ansys/fluent/core/launcher/pim_launcher.py`: Added `ui_mode` as an
optional parameter in the constructor, allowing users to define the user
interface mode for Fluent.
* `def __init__` in `src/ansys/fluent/core/launcher/pim_launcher.py`:
Updated the docstring to include a description of the new `ui_mode`
parameter, which corresponds to values in the `UIMode` enum.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>