docs: Update a setting radiation headlamp [skip tests] (#4334)
This pull request updates the `radiation_headlamp.py` example to improve
the configuration of radiation settings for clarity and flexibility. The
most important change involves modifying the `direct_irradiation`
settings to use a more descriptive and structured format.

### Improvements to radiation settings:

*
[`examples/00-fluent/radiation_headlamp.py`](diffhunk://#diff-98ea9e6f31bfe4fae0fdbac7d8516f05ac6fd2cd1ad8e15278b861a44581c428L473-R475):
Updated the `direct_irradiation` settings to replace the previous
shorthand key `"s-"` with a more descriptive `"Full-spectrum"` key. The
new format includes an additional layer of structure with `"option"` and
`"value"` fields, making the configuration more explicit and extensible.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>