feat: Support named expressions as input to reductions functions. (#4536)
## Context
Named expressions were not supported as an input for the "expression" in
reductions functions.

## Change Summary
Support for named expressions is included.

## Rationale
Named expressions has an attribute "definition", which is checked to get
the value and used.

Sample usage:

```python

>>> absolute_pressure_expression = solver_named_expressions.create()
>>> absolute_pressure_expression.definition = "AbsolutePressure"

>>> solver.fields.reduction.area_average(
>>>         expression=absolute_pressure_expression,
>>>         locations=solver.setup.boundary_conditions.velocity_inlet,
>>>     )
102315.00125
```


## Impact
Reduction functions now will support named expressions in input as well.


Please refer tests for detailed usage.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>