fix: SettingsBase.__call__ should also set state if any argument is passed. (#4551)
## Context
The method `__call__` used to only get the state for parameters. For
some recently modified parameter definitions in settings API, we need to
set the state if any argument is passed to `__call__` to support
backward compatibility. This was already implemented for groups.

## Change Summary
Both groups and parameters should set the state if any argument is
passed to the `__call__` method.

## Rationale
The requirement follows from the settings API.

## Impact
Will fix the doc example `exhaust_system_settings_api` which has been
failing in 26.1.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>