fix: set-state via call (#4567)
## Context
There was an issue in the previous PR that implemented parameter
set-state via `__call__` as the value was not sent to Fluent in the
expected format.

## Change Summary
Only the first argument value is sent to Fluent while setting parameter
state via `__call__`.

## Rationale
To send the set-state value to Fluent in the correct format 

## Impact
New unittest has been added, the example script
`exhaust_system_settings_api.py` will pass.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>