ci: Fail the whole nightly job at the end if the testing step has failed [skip tests] (#4564)
## Context

The testing step has been marked with `continue-on-error: true` in the
nightly testing CIs because we want to execute additional steps on
failure (e.g. notify in Teams). Due to this the failure status is not
being reported at the job level. An example -
https://github.com/ansys/pyfluent/actions/runs/18671795452

<img width="1199" height="766" alt="image"
src="https://github.com/user-attachments/assets/d6ab4f0a-611a-4985-a5e9-daa2c165fa73"
/>

## Change Summary
A final step has been added to set the job as failed if the testing step
was failed before.

## Rationale
We want to mark the job failed in a delayed manner, so that we can
execute additional steps on failure. By default, a job retuns
immediately from a failed step.

## Impact
This change will correctly mark the failed job in GitHub.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>