fix: Fix the slurm-launcher hang by closing stdout after reading (#4611)
## Context
Fluent launched via PyFluent's slurm launcher can hang in some systems
due to stdout pipe buffer overflow. Please check the issue details for
more details.

## Change Summary
PyFluent needs to captures Fluent's stdout to read the slurm job id
printed by Fluent. In this PR, the stdout capture is closed after the
job id is read.

## Rationale
In this approach, PyFluent captures Fluent's stdout till the required
information is read. The resource is freed afterwards which should fix
the buffer overflow issue.

## Impact
The existing behaviour should improve. I'm not able to test it in my
slurm setup as the issue is system-specific. The original requester will
test the fix with a dev package.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>