chore: Update launcher tests for better coverage (#4629)
## Context
Extensive test coverage for the launcher APIs from session classes is
introduced.

One issue is when a 'PureMeshing' session is launched from PyFluent and
'Meshing' session is connected to it using the launcher api, it might be
confusing for users, as now no mechanism exists in server to inform
client that it has been launched in 'PureMeshing' mode. **Hence, this
limitation is to be documented.**

## Change Summary
Tests were added for extensive coverage of launcher APIs and the
limitation is documented.

## Impact
No direct impact, only a limitation as mentioned above to be documented.



**There are four possible connection patterns:**

Meshing → Meshing | ✔ Safe | Full capabilities preserved

PureMeshing → PureMeshing | ✔ Safe | Capabilities match exactly

Meshing → PureMeshing | ✔ Allowed | API connects to a subset of
available capabilities

PureMeshing → Meshing | ⚠ Not recommended | Meshing API expects solver
features that a pure-meshing session cannot provide

The fourth scenario is problematic because Meshing may assume
capabilities (e.g., solver mode switching) that are not enabled in a
pure-meshing session, leading to inconsistent or undefined behavior.

In later Fluent release we might consider updating the app-utilities, as
suggested by @mkundu1

"An implementation can be done by adding a new method
set_pure_meshing_app_mode in the app_utilities service. The new method
will be called to set a flag in Fluent when the Fluent session is
launched or connected from a PureMeshing PyFluent session. That flag
will be checked in get_app_modemethod of app_utilities. We can also
block switch-to-solver operation within Fluent when that flag is set."

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>