feat: Make sure slurm launcher works even if Slurm is not available in client (#4713)
## Context
In the original design of slurm launcher it was assumed that the client
machine is also within the same cluster where Slurm for Fluent is
running and a few client-side features for querying/controlling Slurm
jobs were added based on that assumption. Although these client-side
features provide usability, they are not necessary for the basic
functionality of PyFluent's Slurm launcher: launch and connect to a
Fluent session within a Slurm environment. OTOH, it is a common use case
that PyFluent is running outside a Slurm cluster where user wants to
launch Fluent from PyFluent.

## Change Summary
In this PR, the strict requirement of Slurm availability on the client
machine has been revoked. If Slurm is available in the client machine,
those client-side features for job query/control will work. Otherwise,
they will throw a RuntimeError. A config var is also added to avoid
using Slurm in the client machine even when it is available considering
the scenario where the client machine is part of a different Slurm
cluster than where Fluent would be run.

## Rationale
The basic functionality will be available irrespective of the client
machine being part of Fluent's Slurm cluster. If the client machine is
in Fluent's Slurm cluster, additional usability feature will be
available.

## Impact
It will be possible to use the Slurm launcher from outside a Slurm
environment.

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>